Make RtpSession #write size agnostic.
The #pkt_to_udp_payload method would only allow to write a total of 1500
bytes given the hard-coded size of the internal fixed-lenght u8 array
used to store the bytes.

Change the method to use a dynamic array structure instead (Vec), so it
is able to allocate dynamically the space needed according to the
payload passed in the RtpPkt to write.  The return value is also no
longer a fixed-sized array limited to 1500 bytes, but a dynamic Vec.

The #read method in RtpSession was also slightly modified to trim the
slice that is passed to #parse_pkt, which parses the payload and
transforms it into an RtpPkt. Since the initial array is being
initialised with zero, this means no more unnecessary trailling zeros.