Add "send only" test.
In order to test if sending RTP packets is sucessful (mostly by sending
to a receiving oRTP instance), a new test was added, named test_sending.

It first initialises an RtpSession, then opens a local file and, in a
while loop, while there's something to read, sends the payload read into
the opened RtpSession.

The test_connect was now renamed to the more appropriate name
test_receiving