Provide RtpSession with interior mutability.
Having the write and read methods of the RtpSession require a mutable
reference of self makes it not as usable, since the caller will need to
make sure it has a mutable reference in order to read or write and RTP
packet, which, when concurrency exists, may lead to the caller having to
lock around these operations.

To avoid this, the RtpSession operations, like reaed and write, should
have interior mutability and only require a clean reference to self
(immutable) in order to perform their operations.

The read / write methods of the RtpSession were changed in order to
accept a self (immutable) reference.
Some methods from RtcpStream (called from the RtpSession) were also
requiring mutable self references, thus some member fields in RtpStream
were changed in order to provide interior mutability, by wrapping them
around an Arc<Mutex<>>; The respective methods were then changed to deal
with the wrapping.

The tests were also updted to reflect this usage.