Fix parsing of payload into an RtpPkt.
The payload considered at parse_pkt was being cut both at the beginning
and end, while trying to parse it into a proper RtpPkt.

The end was being cut by a byte, since the only payload.len()-1 bytes
were being copied, while the beginning of the payload was also being
considered to be later than it should, because of a logic error (the
header size was being added twice.