Add support for integration with ICE.
This commit adds rustun as a new dependency. This STUN implementation in
Rust is now used to support integration with ICE. This is achieved by
instantiating a STUN server instance for each new RTP session, and
providing it with a handler (in handlers.rs). Rustun will listen for
STUN requests and use the handler to check if a message is a STUN
request with the expected parameters, or otherwise an RTP request. In
case of being a STUN request, it formats the answer appropriately and
replies. Otherwise, if it is an RTP request, it feeds it back into our
RtpSession.

For communication between rustun and rir there's a new Transport trait
(which, for now, its only implementation is UdpTransport) and a
StunWrapper struct. The latter owns a mspc channel, which is used to
pass the RTP messages from the handler back into the RtpSession, and a
Transport.

Another new trait has also been introduced, RirHandler, which clients
can implement to pass handlers into rir, in order to get events back.