Give interior mutability to RtpSession.
Previously, callers of the RtpSession API would need to be aware of
mutability in order to make calls to methods like change_transport() and
change_rtcp_transport(), since it would expect a mutable instance.

The behaviour has now changed and callers shouldn't now worry about
mutability as that is dealt inetrnally. For that, certain fields are now
wrapped around either a Mutex (sender of StunWrapper) or RwLock's
(local_socket and rem_socket in UdpTransport).