stun: Implement StunPkt and Attr help functions.
Previously, if one wanted to get a Username attribute from a parsed
packet, for example, one would do:
    let username = pkt.attrs.get(&0x0006);
    if let &Attr::Username( ref x ) = username.unwrap() {
        ...
    }

This would lead to duplicate code when trying to get all attributes
present in a packet.

Instead, this has now been modified so a StunPkt is implemented to
support functions such as get_username(), which get a username from the
internal `attrs` member and calls the struct Attr's own functions, in
this case username(), to convert the Attr instance in a Username
instance.

Thus, one can now do:
    let username = pkt.get_username();
    if let Some(x) = username {
        ...
    }