stun: Move parse_attr() core match to Attr impl.
parse_attr() was aware of the types of attributes, like 0x0006 for a
Username attribute, and was responsible for mapping that attribute type
to a specific Attr. This isn't ideal as it means the logic to parse a
raw attribute is not entirely encapsulated in the Attr impl.

Instead, the match has been moved to a from_raw() new function in Attr
impl, which means the logic to parse a raw attribute to an Attr struct
is now encapsulated in the Attr impl.