stun: Modify struct RawAttr to store references.
This modification allows to save a copy and instead a reference to the
slices of the raw attribute (attr_raw) and the raw message preceeding
the attribute (precd_msg) is kept. Later, in attributes which need to
store the precd_msg, such as for MessageIntegrity and Fingerprint, this
is copied to a Vec.

As a consequence, the signature of the from_raw() function in the Attr
impl has been changed to take only the attribute type and RawAttr, since
all the information needed is now calculated a priori on the caller and
passed on through the RawAttr.

Also, passwd member has been drop as there were no users.