stun: Split Stun struct impl validate().
Instead of verifying the FINGERPRINT in validate(), a is_valid()
function has now been added to StunPkt struct impl which checks
unconditionally for the FINGERPRINT attribbute. This is because rfc5245
defines FINGERPRINT presence as mandatory.

This changes validate() to be responsible for performing the later
stages of the STUN validation, which is, performing the checks for the
USERNAME and MESSAGE-INTEGRITY attributes and authenticating the STUN
messages, as well as gathering any unknown but required attributes. It
now also returns a Result, instead of an Option.