stun: Don't authenticate Indications.
rfc5245 explicity mentions that Binding Indications shouldn't use any
other attribute than the FINGERPRINT attribute. Thus, process_stun() has
been changed so the authentication validation is not called for
Indication messages, and only the unknown but required attributes are
validated.

For this, the validate() function was split in two other, authenticate()
and val_unknown_attrs, which perform the authentication and validate the
unknown but required attributes, respectively.

Additionaly, in order to flag if the message was indeed a STUN message
or not, the return type of the process_stun() has been changed to a
(bool, Option). The new parameter, the bool parameter, can be looked at
as a "complete" flag, which indicates if the message has been fully
processed or if there was some error and the message should continue
processing as a different protocol (if being multiplexed). The only
caller, in rtp.rs has been modified to include this logic.