stun: Do not modify header length in attrs.
Some attributes' to_raw(), such as FINGERPRINT and MESSAGE-INTEGRITY,
need to (temporarily) modify the header length so the CRC and hash are
calculated with the correct header set. However, the to_raw()
implementation of each attribute is only supposed to return the raw set
of bytes that represent the attribute, and not change the message's
content.

This changes the to_raw() in these two attributes to re-set the previous
header length, after the calculations are done.