stun: Reorder attributes in StunPkt impl to_raw().
The MESSAGE-INTEGRITY and FINGERPRINT attribute must be the second to
last and last attributes in a StunPkt message, respectively. However, in
messages with an Error attribute, they were being inserted before the
error, causing authentication and validation errors. It has been fixed
so they are inserted in the correct order.