stun: Fix padding in UnknownAttrs' to_raw().
to_raw()'s implementation of UnknownAttrs wasn't padding the attribute
in a case of an odd number of unknown attributes was given, which could
cause messages to fail to validate.

Also, this turns to_raw() in a method, as it happens for most of other
attributes to_raw() implementations already.