stun: Enforce the Username attribute.
Up until now the Username attribute wasn't being enforced, as it wasn't
being validated against anything.

Instead, a Stun instance now has a pair of credentials, a username and a
password - to represent the agent's credentials. The Username attribute
is then validated against the local username. A new method has been
added to the Username struct impl, match_on_username(), which, given an
input username, performs the validation against the Username attribute,
returning 'true' or 'false'. This new method is called when
authenticate()'ing the parsed StunPkt.

Also, the rtp.rs module has been modified in order to reflect the above
changes, which means that a Stun::new() now needs to be provided with a
pair of username and password.