rtp: Fix RTCP Sender Report parse.
While parsing an RTCP Sender Report packet, in parse_rtcp_pkt(), the
"index" wasn't being correctly calculated and updated, causing the parse
to panic. Additionally, not only the calls to read a u32 BigEndian were
being called with the wrong "index", but were also being given the wrong
slice. Instead, the "index" is now updated as the payload is parsed and
read, and the calls to read a u32 BigEndian are given an appropriate 4
byte slice.

This was harder to debug than it should be because the thread was
blowing up without any stack trace, while the main process would
continue running. The consequence was that STUN responses would stop
being sent by the RTCP thread which would cause the other side to
disconnect and consider the respective ICE candidate "failed".