rtp: Abstract RtpSession and RtcpStream transport.
Although the StunWrapper implementation is already abstracted by the
Transport trait, both RtpSession and RtcpStream were using the
StunWrapper implementation directly. Instead, RtpSession and RtcpStream
now use a Transport object trait, which could be either a UdpTransport
or a StunWrapper implementation (or more in the future).

A later commit will make use of this to introduce a new API, so a simple
connection can be set up with no STUN involved, and thus using directly
the UdpTransport implementation.

The Transport trait has also been extended with a new method,
set_remote(), to enable to caller to set the peer's remote address.